Improved factorization signification
Rather than compute a list of primes from 2..n in advance, we can simply walk
our way up from 2 to n. Every time we find a prime number that is a factor of n,
we push the prime into the factors vector, divide n by the prime, and then
repeat. This avoids having to do a lot of work and since this is prime
factorization, most of our numbers _should_ be small.